requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
cssselect==1.2.0
cssutils==2.9.0
disposable-email-domains==0.0.96
distlib==0.3.8
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
emails==0.6
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
lxml==5.1.0
Mako==1.3.2
MarkupSafe==2.1.5
neo4j-driver==5.17.0
packaging==23.2
passlib==1.7.4
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.4.0
premailer==3.10.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlmodel==0.0.14
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
